chore: split getRequester
