+++ lib/node.ts
@@ -1,6 +1,6 @@
 import AlgoliaAnalytics from "./insights";
 import { getFunctionalInterface } from "./_getFunctionalInterface";
-import { getRequesterForNode } from "./utils/request";
+import { getRequesterForNode } from "./utils/getRequesterForNode";
 
 const requestFn = getRequesterForNode();
 const instance = new AlgoliaAnalytics({ requestFn });
